* OFS Plugin: Added the possibility to set the default values for the "adjust" settings. * OFS Plugin: Added the possibility to reset all the "adjust" settings to the default. * OFS Plugin: Added the possibility to automatically reset some of the "adjust" settings when creating virtual points (i.e. the checked "R" will be reset). * OFS Plugin: Minimum Percentage can now be lower then 1. * OFS Plugin: Fixed the channel locking mecanism. You should be able to use the plugin in multiple OFS opened at the same time.

+++ src/FunscriptToolbox.Core/MotionVectors/FrameAnalyser.cs
@@ -21,13 +21,13 @@ namespace FunscriptToolbox.Core.MotionVectors
             this.QualityLevel = qualityLevel;
         }
 
-        public FrameAnalyser Filter(int activityLevel, int qualityLevel, int minPercentage)
+        public FrameAnalyser Filter(int activityLevel, int qualityLevel, double minPercentage)
         {
             var rules = this.Rules
                     .Where(rule => rule.Activity >= activityLevel)
                     .Where(rule => rule.Quality >= qualityLevel)
                     .ToArray();
-            var minRules = this.NbBlocX * this.NbBlocY * minPercentage / 100;
+            var minRules = (int)(this.NbBlocX * this.NbBlocY * minPercentage / 100);
             if (rules.Length < minRules)
             {
                 rules = this.Rules
