+++ src/FunscriptToolbox.Core/MotionVectors/PluginMessages/SharedConfig.cs
@@ -6,7 +6,7 @@
         public double DefaultLearningDurationInSeconds { get; set; }
         public int DefaultActivityFilter { get; set; }
         public int DefaultQualityFilter { get; set; }
-        public int DefaultMinimumPercentageFilter { get; set; }
+        public double DefaultMinimumPercentageFilter { get; set; }
         public double MaximumNbStrokesDetectedPerSecond { get; set; }
         public double MaximumLearningDurationInSeconds { get; set; }
         public bool TopMostUI { get; set; }
